refactor(search): search libs
